[CHANGED FILE/DIR: init.lua, plugin/commands.lua, NEW FILE/DIR: commands/**] Big architecture update: commands are split across several files, so it's easier to add new ones
