+++ init.lua
@@ -1,7 +1,13 @@
 local M = {}
 
+local general = require("LazyDeveloperHelper.commands.general")
+local install = require("LazyDeveloperHelper.commands.install")
+local requirements = require("LazyDeveloperHelper.commands.install_requirements")
+
 function M.setup()
-  require("LazyDeveloperHelper.plugin.commands").commands()
+  general.register()
+  install.register()
+  requirements.register()
 end
 
 return M
